refactor(cli): move _firefox_containers into cli/ package
- git mv src/notebooklm/_firefox_containers.py -> src/notebooklm/cli/_firefox_containers.py
- Update cli/session.py import + docstring
- Update test mock paths (test_firefox_containers.py, cli/test_session.py)
- Update docs/auth-keepalive.md reference
- Boundary plan: .sisyphus/plans/private-module-boundary.md (PR-B)
