+++ src/notebooklm/cli/_firefox_containers.py
